requirements.txt
# backend requirements

fastapi 	#or flask Backend framework (choose one)
uvicorn	 #ASGI server for FastAPI
gunicorn	 #Production-ready WSGI server for Flask
requests	 #Handle API requests
pydantic	 #Define API request/response models

# Data Preparation and RAG pipeline
pandas	 #Handle structured text data
numpy	 #Numerical operations
faiss-cpu	 #Vector search for document retrieval
chromadb	 #Alternative vector database
sentence-transformers	 #Convert text into embeddings
langchain	 #Orchestrate RAG pipeline
PyMuPDF #for pdf file handling

# Knowledge Graph and Visualization
networkx	#Create knowledge graphs
matplotlib	 #Generates static graph visualizations
neo4j	 #Stores and retrieves graph data (optional)

google.generativeai>=0.3.0

# Add requests package
requests>=2.26.0

python-docx>=0.8.11

spacy>=3.0.0
# python -m spacy download en_core_web_sm

transformers>=4.0.0
torch>=1.9.0
networkx>=2.6.0
sentencepiece>=0.1.96  # Required for BART model


